Fix reading data of large SunSpec models
Up until now, the data of a SunSpec model was read as one block. This
does not work for large SunSpec models since the max. number of
registers that can be read as a block via Modbus is 125. The data of
SunSpec models is now read in chunks of max. 125 registers.